requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
openpyxl==3.1.2
python-dotenv==1.0.0
werkzeug==3.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
joblib==1.3.2
